+++ app/uk/gov/hmrc/plasticpackagingtax/registration/views/organisation/partner_type.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import uk.gov.hmrc.plasticpackagingtax.registration.{controllers => pptControllers}
-@import uk.gov.hmrc.plasticpackagingtax.registration.controllers.organisation.{routes => pptRoutes}
+@import uk.gov.hmrc.plasticpackagingtax.registration.controllers.partner.{routes => partnerRoutes}
 @import uk.gov.hmrc.plasticpackagingtax.registration.views.html.main_template
 @import uk.gov.hmrc.plasticpackagingtax.registration.views.model.{BackButton, Title}
 @import uk.gov.hmrc.plasticpackagingtax.registration.forms.organisation.PartnerType
@@ -45,7 +45,7 @@
 
         @pageHeading(messages("nominated.partner.type.title"))
 
-        @formHelper(action = pptRoutes.NominatedPartnerTypeController.submit(), 'autoComplete -> "off") {
+        @formHelper(action = partnerRoutes.NominatedPartnerTypeController.submit(), 'autoComplete -> "off") {
 
             @govukRadios(Radios(
                 fieldset = Some(Fieldset(
