+++ app/uk/gov/hmrc/plasticpackagingtax/registration/views/task_list_partnership.scala.html
@@ -16,7 +16,7 @@
 
 @import uk.gov.hmrc.plasticpackagingtax.registration.controllers.{routes => pptRoutes}
 @import uk.gov.hmrc.plasticpackagingtax.registration.controllers.organisation.{routes => organisationRoutes}
-@import uk.gov.hmrc.plasticpackagingtax.registration.controllers.contact.{routes => contactRoutes}
+@import uk.gov.hmrc.plasticpackagingtax.registration.controllers.partner.{routes => partnerRoutes}
 @import uk.gov.hmrc.plasticpackagingtax.registration.controllers.partnership.{routes => partnershipRoutes}
 @import uk.gov.hmrc.plasticpackagingtax.registration.models.registration.Registration
 @import uk.gov.hmrc.plasticpackagingtax.registration.models.request.JourneyRequest
@@ -74,7 +74,7 @@
         @task(3, "registrationPage.task.contactDetails.heading.partnership", Set(
             TaskSection(
                 title = "registrationPage.task.contactDetails.partnership",
-                link = organisationRoutes.NominatedPartnerTypeController.displayPage(),
+                link = partnerRoutes.NominatedPartnerTypeController.displayPage(),
                 showLink = registration.isCompanyDetailsComplete && registration.isLiabilityDetailsComplete,
                 status = registration.nominatedPartnerDetailsStatus
             ),
